dro_hetero_anchors/requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
pyyaml>=6.0
rich>=13.4
tqdm>=4.66
tensorboard>=2.16
